requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3